requirements.txt
# 核心数据科学库
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
opencv-python==4.8.1
Pillow==10.0.0
matplotlib==3.7.2

# 代码质量和测试
black==23.9.1
ruff==0.1.6
pytest==7.4.3
pytest-cov==4.1.0
pre-commit==3.5.0

# 配置管理
python-dotenv==1.0.0
pydantic==2.4.2

# ML工具
mlflow==2.9.2
dvc==3.32.0
dagshub==0.3.9

# 其他
jupyter==1.0.0